requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-extensions==3.1.5
django-import-export==2.6.1
django-preferences==1.0.0
django-redis==5.0.0
django-simpleui==2021.11.5
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==4.0.1
h11==0.13.0
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
nltk==3.7
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.2
pyactiveresource==2.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.4.3
ShopifyAPI==12.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.3
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
